Simplified the json formatting for more speed at the price of human-readability. The idea is to make the two formats coexist, one for small collection usage and initialisation, the other for exporting and importing back a large collection.
Changed the streaming behaviour of the taggings getters (get-tags-of-media and get-medias-with-tag to a single response package as it is never very large so streaming induces performance losses in this case.

Changed the way all of the sql queries were stated and executed for better string handling

Force cast to str on value for tag creation

Tried to implement multithreading on the FileHandler, but it generates errors from the server side. TBC